start tracking jobseeker update and jobseeker activity

+++ src/main/kotlin/fr/audithor/services/JobSeekerService.kt
@@ -14,7 +14,7 @@ class JobSeekerService(private val jobSeekerRepository: JobSeekerRepository) {
     return jobSeekers.map { DtoMapper.toDto(it) }
   }
 
-  fun getJobSeekerById(id: Int): JobSeeker {
-    return jobSeekerRepository.findById(id.toLong())
+  fun getJobSeekerById(id: Long): JobSeeker {
+    return jobSeekerRepository.findById(id)
   }
 }
